* Fixed the last of the Windows warnings * New C_rand_real() function returns a random number between 0-1. * Windows project file debug mode fixed * C_print() no longer causes the output to be wrapped * Fixed Windows compatibility (strncasecmp)

+++ src/common/c_string.c
@@ -508,7 +508,7 @@ int C_utf8_strlen(const char *str, int *chars)
  bytes. Returns the size of the UTF-8 character or 0 if there is not enough
  room or if [src] starts with NUL.
 \******************************************************************************/
-int C_utf8_append(char *dest, int *dest_i, size_t dest_sz, const char *src)
+int C_utf8_append(char *dest, int *dest_i, int dest_sz, const char *src)
 {
         int len, char_len;
